uhh idk not working yet:

+++ student-distrib/syscalls.h
@@ -101,8 +101,8 @@ typedef struct pcb {
 
 
 /* Forward declarations */
-int32_t halt(uint32_t status);
-int32_t wrapper_halt(uint8_t status);
+int32_t halt(uint8_t status);
+int32_t wrapper_halt(uint32_t status);
 int32_t execute (const uint8_t* command);
 int32_t read (int32_t fd, void* buf, int32_t nbytes);
 int32_t write (int32_t fd, const void* buf, int32_t nbytes);
